Remove collector image
The CI image for collector used to be used on CircleCI and OSCI, but
once we migrated to GHA we stopped using it, since the base VMs already
have all the tools we need pre-installed in them.

Removing the image so colleagues using this repo no longer need to worry
about maintaining it.