Add PG flavor to the stackrox-test image
stackrox-test image is used to run go-postgres tests and has PostgreSQL
installed for that purposes. Now we come to the point that ACS has to be
tested against multiple versions. To support that, add all needed
binaries into the test image and let the consumers choose which version
to use via modifying the PATH.